requirements.txt
Flask==3.0.0
supabase==2.3.0
python-dotenv==1.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
Werkzeug==3.0.1
cryptography==41.0.7